# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: market
    description: EU market inputs and forecast operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /market/overview:
    get:
      operationId: getMarketOverview
      tags: [market]
      summary: Get current EU steel market inputs
      description: Returns the latest available inputs for a selected EU country, with freshness and source metadata.
      parameters:
        - name: country
          in: query
          required: false
          schema:
            type: string
            default: Germany
            enum: [Germany, France, Italy, Poland, Spain, Netherlands, Belgium]
      responses:
        "200":
          description: Market overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketOverview"
  /market/forecast:
    get:
      operationId: getMarketForecast
      tags: [market]
      summary: Get a directional production cost forecast
      description: Returns forecast points with an uncertainty band that widens with the horizon.
      parameters:
        - name: country
          in: query
          required: false
          schema:
            type: string
            default: Germany
            enum: [Germany, France, Italy, Poland, Spain, Netherlands, Belgium]
        - name: horizon
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 26
            default: 26
      responses:
        "200":
          description: Forecast series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketForecast"
  /market/validation:
    get:
      operationId: getMarketValidation
      tags: [market]
      summary: Get per-series model validation metrics
      description: Returns time-based held-out validation metrics for HRC, electricity, gas, and EUA models.
      parameters:
        - name: country
          in: query
          required: false
          schema:
            type: string
            default: Germany
            enum: [Germany, France, Italy, Poland, Spain, Netherlands, Belgium]
      responses:
        "200":
          description: Per-series validation metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketValidation"
  /market/backtest:
    get:
      operationId: getMarketBacktest
      tags: [market]
      summary: Get rolling forecast backtest metrics
      description: Returns error metrics calculated only from matured forecast snapshots and complete historical source observations.
      parameters:
        - name: country
          in: query
          required: false
          schema:
            type: string
            default: Germany
            enum: [Germany, France, Italy, Poland, Spain, Netherlands, Belgium]
      responses:
        "200":
          description: Historical forecast accuracy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketBacktest"
  /market/assumptions:
    get:
      operationId: getMarketAssumptions
      tags: [market]
      summary: Get model assumptions and source notes
      responses:
        "200":
          description: Model assumptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketAssumptions"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Freshness:
      type: string
      enum: [live, cached, estimated]
    MarketInput:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        value:
          type: number
        unit:
          type: string
        freshness:
          $ref: "#/components/schemas/Freshness"
        source:
          type: string
        statusMessage:
          type: string
        provenanceKind:
          type: string
          enum: [official, licensed_benchmark, proxy, assumption]
        provenanceNote:
          type: string
        updatedAt:
          type: string
        lastFetchedAt:
          type: string
        sourceRefreshInterval:
          type: string
          enum: [daily, weekly, monthly]
        nextExpectedUpdate:
          type: string
      required: [key, label, value, unit, freshness, source, statusMessage, provenanceKind, provenanceNote, updatedAt, lastFetchedAt, sourceRefreshInterval, nextExpectedUpdate]
    CountryAdjustment:
      type: object
      properties:
        country:
          type: string
        electricityMultiplier:
          type: number
        laborMultiplier:
          type: number
        freightMultiplier:
          type: number
        subsidyNote:
          type: string
      required: [country, electricityMultiplier, laborMultiplier, freightMultiplier, subsidyNote]
    MarketOverview:
      type: object
      properties:
        country:
          type: string
        asOf:
          type: string
        inputs:
          type: array
          items:
            $ref: "#/components/schemas/MarketInput"
        adjustment:
          $ref: "#/components/schemas/CountryAdjustment"
        baseCostPerTon:
          type: number
        confidenceScore:
          type: number
        liveInputCount:
          type: number
        totalInputCount:
          type: number
      required: [country, asOf, inputs, adjustment, baseCostPerTon, confidenceScore, liveInputCount, totalInputCount]
    ForecastPoint:
      type: object
      properties:
        week:
          type: number
        label:
          type: string
        costPerTon:
          type: number
        lower:
          type: number
        upper:
          type: number
      required: [week, label, costPerTon, lower, upper]
    MarketForecast:
      type: object
      properties:
        country:
          type: string
        horizon:
          type: number
        points:
          type: array
          items:
            $ref: "#/components/schemas/ForecastPoint"
        series:
          type: array
          items:
            $ref: "#/components/schemas/SeriesForecast"
        validation:
          $ref: "#/components/schemas/MarketValidation"
        backtest:
          $ref: "#/components/schemas/MarketBacktest"
        methodology:
          type: string
      required: [country, horizon, points, series, validation, backtest, methodology]
    SeriesHistoryPoint:
      type: object
      properties:
        week:
          type: number
        label:
          type: string
        value:
          type: number
      required: [week, label, value]
    SeriesForecastPoint:
      type: object
      properties:
        week:
          type: number
        label:
          type: string
        value:
          type: number
        lower:
          type: number
        upper:
          type: number
      required: [week, label, value, lower, upper]
    SeriesForecast:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        unit:
          type: string
        model:
          type: string
        sourceInputKey:
          type: string
        provenanceKind:
          type: string
          enum: [official, licensed_benchmark, proxy, assumption]
        history:
          type: array
          items:
            $ref: "#/components/schemas/SeriesHistoryPoint"
        points:
          type: array
          items:
            $ref: "#/components/schemas/SeriesForecastPoint"
      required: [key, label, unit, model, sourceInputKey, provenanceKind, history, points]
    ValidationRow:
      type: object
      properties:
        series:
          type: string
        model:
          type: string
        baseline:
          type: string
        mae:
          type: number
        rmse:
          type: number
        mape:
          type: number
        baselineMape:
          type: number
        vsBaseline:
          type: number
        sampleSize:
          type: number
        validationWindow:
          type: string
      required: [series, model, baseline, mae, rmse, mape, baselineMape, vsBaseline, sampleSize, validationWindow]
    MarketValidation:
      type: object
      properties:
        generatedAt:
          type: string
        confidenceScore:
          type: number
        freshnessCoverage:
          type: number
        liveSeriesCount:
          type: number
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ValidationRow"
        methodology:
          type: string
      required: [generatedAt, confidenceScore, freshnessCoverage, liveSeriesCount, rows, methodology]
    BacktestWindow:
      type: object
      properties:
        windowDays:
          type: number
        windowStart:
          type: string
        windowEnd:
          type: string
        observationCount:
          type: number
        meanAbsolutePercentageError:
          type: ["number", "null"]
        medianAbsolutePercentageError:
          type: ["number", "null"]
        bandCoveragePercent:
          type: ["number", "null"]
        status:
          type: string
          enum: [ready, insufficient]
      required: [windowDays, windowStart, windowEnd, observationCount, meanAbsolutePercentageError, medianAbsolutePercentageError, bandCoveragePercent, status]
    MarketBacktest:
      type: object
      properties:
        generatedAt:
          type: string
        sampleWindow:
          type: string
        observationCount:
          type: number
        rolling30:
          $ref: "#/components/schemas/BacktestWindow"
        rolling90:
          $ref: "#/components/schemas/BacktestWindow"
        errorBandMethodology:
          type: string
      required: [generatedAt, sampleWindow, observationCount, rolling30, rolling90, errorBandMethodology]
    AssumptionItem:
      type: object
      properties:
        label:
          type: string
        detail:
          type: string
        status:
          $ref: "#/components/schemas/Freshness"
        refresh:
          type: string
      required: [label, detail, status, refresh]
    MarketAssumptions:
      type: object
      properties:
        title:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/AssumptionItem"
        disclaimer:
          type: string
      required: [title, items, disclaimer]