# lib/api-zod/src/generated/types/marketInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Freshness } from './freshness';
import type { MarketInputProvenanceKind } from './marketInputProvenanceKind';
import type { MarketInputSourceRefreshInterval } from './marketInputSourceRefreshInterval';

export interface MarketInput {
  key: string;
  label: string;
  value: number;
  unit: string;
  freshness: Freshness;
  source: string;
  statusMessage: string;
  provenanceKind: MarketInputProvenanceKind;
  provenanceNote: string;
  updatedAt: string;
  lastFetchedAt: string;
  sourceRefreshInterval: MarketInputSourceRefreshInterval;
  nextExpectedUpdate: string;
}